// src/DockiUp.Frontend/src/app/api/model/composeValidationDto.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ComposeServiceDto } from './composeServiceDto';


export interface ComposeValidationDto { 
    valid: boolean;
    errors: Array<string>;
    warnings: Array<string>;
    services: Array<ComposeServiceDto>;
}