// src/DockiUp.Frontend/src/app/api/model/repositoryInspectionDto.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RepositoryComposeFileDto } from './repositoryComposeFileDto';


export interface RepositoryInspectionDto { 
    branch: string;
    defaultBranch: string | null;
    branches: Array<string>;
    composeFiles: Array<RepositoryComposeFileDto>;
    defaultComposeFile: string | null;
}